feat: implement POC
