+++ package.json
@@ -35,7 +35,8 @@
 	"dependencies": {
 		"@actions/core": "^1.10.1",
 		"@actions/github": "^6.0.0",
-		"zod": "^3.22.4"
+		"zod": "^3.22.4",
+		"zod-validation-error": "^2.1.0"
 	},
 	"devDependencies": {
 		"@types/eslint": "^8.56.0",
@@ -57,8 +58,7 @@
 		"rimraf": "^5.0.5",
 		"tsup": "^8.0.1",
 		"typescript": "^5.3.3",
-		"vitest": "^1.1.1",
-		"zod-validation-error": "^2.1.0"
+		"vitest": "^1.1.1"
 	},
 	"packageManager": "pnpm@8.7.0",
 	"engines": {
