--- src/types.ts
@@ -1,5 +1,0 @@
-export interface GreetOptions {
-	logger?: (message: string) => void;
-	message: string;
-	times?: number;
-}
